+++ NEWS.md
@@ -1,6 +1,8 @@
-# lipdR 0.3.6
-
 # lipdR 0.3.5
+
+* new function `removeEnsembles()` that will strip all ensembles from a LiPD object for portability
+* `readLipd()` now allows a vector of paths as input, to create a multiLipd from mulitple directories
+* Progress bar for writeLipd()
 
 # lipdR 0.3.4